Made update , try to detect memory and give memory mapping

+++ BOOT/stage1_boot.c
@@ -1,6 +1,6 @@
 extern void read_sectors(), call_second_boot();
 
-short bootdrive;
+int bootdrive;
 
 int err = 0; // contain 1 if error
